* 'master' of https://bitbucket.org/lei-isep/lapr3-2020-g033:
  AUCMN1-102 #progress
  AUCMN1-102 #progress Update UC17
  AUCMN1-209 #comment removed comments
  no message
  AUCMN1-161 #progress UC15 [Development]
  AUCMN1-209 #comment Added comments to avoid sonar analysis
  AUCMN1-102 #progress Update UC
  AUCMN1-211 #comment changes to MR
  AUCMN1-91 #comment updated the tests
  AUCMN1-90 #comment Updated the ElectricScooter
  AUCMN1-212 #comment changed id to licensePlate + added new functions
  AUCMN1-163 #progress Connection to database -Updates
  AUCMN1-212 #comment changes to tables
  AUCMN1-167 #start-progress Changed the struct of the C function
  AUCMN1-167 #start-progress Started the development on the assembly function
  AUCMN1-161 #problem build solve [Development]
  AUCMN1-161 #start of development of UC15 [Development]
  AUCMN1-212 #comment added new functions and procedures to database

# Conflicts:
#	src/main/java/lapr/project/controller/VehicleController.java
#	src/main/java/lapr/project/data/ParkHandler.java
#	src/main/java/lapr/project/data/ScooterHandler.java

+++ C and Assembly/LAPR3/struct.h
@@ -3,10 +3,10 @@
    typedef struct {
      int scooterID;
      int parkID;
-     double charging_place_potency;
-     double ah_battery;
-     double maxBattery;
-     double actualBattery;
+     int charging_place_potency;
+     int ah_battery;
+     int maxBattery;
+     int actualBattery;
    } s1;
 
 #endif
